Adding data source factory mechanism to enable external data sources to be added

+++ src/map_types/dim_mapping.cpp
@@ -1,6 +1,7 @@
 #include "map_types/dim_mapping.hpp"
 
 #include <cstdlib>
+#include <cstdint>
 
 #include "exceptions/exceptions.hpp"
 #include "map_types/base_mapping.hpp"
@@ -17,5 +18,5 @@ libtokamap::TypedDataArray libtokamap::DimMapping::map(const MapArguments& argum
         throw libtokamap::MappingError{"cannot use DIM_PROBE on rank 0 mapping '" + m_dim_probe + "'"};
     }
     // TODO: Add DIM_INDEX for selecting which dimension to return size of
-    return TypedDataArray{static_cast<int>(array.shape()[0])};
+    return TypedDataArray{static_cast<uint64_t>(array.shape()[0])};
 }
